MakeFile.toml: patch: Check for semver compliance
Per the documentation regarding the [patch] option in the Cargo.toml
file, the only thing that the [patch] option does is add that source
to the list of sources Cargo will search for dependencies for. The
dependency resolver still works the same way. Due to the fact that
the dependency resolver is greedy, it will always choose the highest
available version of a dependency.

To make patching more reliable, we add a check to our patch command to
ensure that the patched crate's version is greater than or equal to
the current version of the crate in the Cargo.lock file. This should
help with the common issue of Cargo choosing a non-patched version of
the crate, even when a patch is applied.