Q35: Add the Patina Performance Component
This adds the Patina Performance component to the Q35 binary so it
is a part of testing and demonstration in the Patina QEMU repo.

Note: Arm support for performance was added recently. I have not
tested that and this is part of unblocking performance integration
on Intel platforms, so that can be added to the SBSA binary in the
future, but it is not done in this change.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>