Q35/SBSA: Enable Debugger By Default in Debug Builds
In order to show a realistic platform example, this commit
enables the debugger by default in debug builds. This ensures
that the file space required for it will always be available
in debug builds. It is disabled by default in release builds
to save the space for production.

In order to enable CLI enabling of the initial breakpoint,
which was previously provided for, another feature flag is added to
enable the initial breakpoint.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>