Update to Patina v12.1.0
Integrates the latest patina release. The main integration change
needed is to update the path for sample components used in the Q35
platform build.

The version of qemu_dxe_core is updated to 2.0.1. Only the patch
version is updated since this release is not expected to require
any specific integration work needed by a platform using the binary.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>