Run patina_tests by Default on Q35 and SBSA
The patina test feature was enabled on these platforms,
but the test runner component wasn't added, so tests weren't
actually getting executed. This adds the test runner to run
all tests. Eventually, this may be filtered.

This also updates Cargo.lock to pull in Patina 17.1.0 which
contains new patina_tests.

Signed-off-by: Oliver Smith-Denny <osde@microsoft.com>